refactor & incorporate error reason from api
